Complete Project
Note to self: really learn jQuery. Whilst it was easy knowing
what to do. It was much harder figuring out HOW TO do it.
In fact, this commmit might not meet the requirements from the
control script.

Most importantly perhaps, the Javascript code is not very
object oriented. The majority of the code, handcraft the HTML
code which is not a good long terms solution.

Perhaps in the next commit I can try to create a new element
in the same fashion as 'newDiv' with the properties and all.

   --- Lionel Saliou, Ph.